refactor(protocol): split the receipts vocabulary out of event.rs (pure move)
event.rs sits at its file-size ceiling (2,636 lines, grandfathered in the
ratchet baseline), and Phase 2 must add fields to the manifest and the frame
reference. Growing a file at its ceiling means raising the ceiling, which the
ratchet exists to make hard; splitting it first is the alternative the guard
is designed to encourage.

The seam is the one already there: BlockKind, CacheZone, ContextFrameRef,
BlockOrigin, ManifestEntry, ProviderShare, ContextProviderUsage and
ContextUsage are the persisted receipt of one model call. They share a
forward-compatibility discipline the rest of the event vocabulary does not
need, and they are the only types in the file that a store reads back rather
than a renderer consuming live.

Verbatim move, no wire change: the types are re-exported from `event` as well
as the crate root, so `stella_protocol::event::BlockKind` and
`stella_protocol::BlockKind` both still resolve. Tests are untouched and stay
where they are — they exercise these types through AgentEvent, which is where
they belong.

Refs #713
